Move cmd dir contents into entrypoint package
