+++ pkg/entrypoint/trace.go
@@ -1,4 +1,4 @@
-package main
+package entrypoint
 
 import (
 	"context"
